few big fixed. code improved. game frame improved.  all literal'll be constant-fy

+++ src/puyoResources/puyoPrinting/printObject/gamePage/puyoPrintScreenhead.cpp
@@ -8,5 +8,5 @@ using namespace sf;
 puyoPrintScreenhead::puyoPrintScreenhead(Sprite screen, int life) : puyoPrintObject(screen,life)
 {
     const auto [ignore,player_board_y] = PLAYER_BOARD_POS[0];
-    sprite.setTextureRect(IntRect({0,0}, {static_cast<int>(SCREEN_X),player_board_y+1})); 
+    sprite.setTextureRect(IntRect({0,0}, {(int)(SCREEN_X),(int)player_board_y+1})); 
 }
